Add core modules: attendance, auth, and user settings with DTOs and exporters

+++ src/common/entity/leave-class.entity.ts
@@ -8,7 +8,7 @@ export class LeaveClass {
   @Column({ name: 'LeaveName', type: 'varchar', length: 20 })
   leaveName: string;
 
-  @Column({ name: 'MinUnit', type: 'double', default: 1 })
+  @Column({ name: 'MinUnit', type: 'float', default: 1 })
   minUnit: number;
 
   @Column({ name: 'Unit', type: 'smallint', default: 1 })
@@ -23,7 +23,7 @@ export class LeaveClass {
   @Column({ name: 'ReportSymbol', type: 'varchar', length: 4, default: '-' })
   reportSymbol: string;
 
-  @Column({ name: 'Deduct', type: 'double', default: 0 })
+  @Column({ name: 'Deduct', type: 'float', default: 0 })
   deduct: number;
 
   @Column({ name: 'Color', type: 'int', default: 0 })
